+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-25T05:49:08.698Z",
+    "2026-05-25T07:24:02.374Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {